[feature] 로그인 -> 프로젝트 목록 -> 해당 프로젝트 로직 변경 및 디자인 수정

+++ src/services/aiOpenAnalysis/app/proEnter/ProEnterTab1.jsx
@@ -94,7 +94,7 @@ const ProEnterTab1 = (props) => {
                     return;
                 }
                 if (res?.success === "777") {
-                    modal.showAlert("알림", "등록이 완료되었습니다.", {
+                    modal.showConfirm("알림", "등록이 완료되었습니다.", {
                         btns: [{
                             title: "확인", click: () => {
                                 sessionStorage.setItem("projectnum", "");
